+++ apps/connect/source/ftrack_connect/utils/directory.py
@@ -1,0 +1,47 @@
+# :coding: utf-8
+# :copyright: Copyright (c) 2014-2023 ftrack
+
+import os
+import subprocess
+import sys
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+def open_directory(path):
+    '''Open a filesystem directory from *path* in the OS file browser.
+
+    If *path* is a file, the parent directory will be opened. Depending on OS
+    support the file will be pre-selected.
+
+    .. note::
+
+        This function does not support file sequence expressions. The path must
+        be either an existing file or directory that is valid on the current
+        platform.
+
+    '''
+    if os.path.isfile(path):
+        directory = os.path.dirname(path)
+    else:
+        directory = path
+
+    if sys.platform == 'win32':
+        # In order to support directories with spaces, the start command
+        # requires two quoted args, the first is the shell title, and
+        # the second is the directory to open in. Using string formatting
+        # here avoids the auto-escaping that python introduces, which
+        # seems to fail...
+        subprocess.Popen('start "" "{0}"'.format(directory), shell=True)
+
+    elif sys.platform == 'darwin':
+        if os.path.isfile(path):
+            # File exists and can be opened with a selection.
+            subprocess.Popen(['open', '-R', path])
+
+        else:
+            subprocess.Popen(['open', directory])
+
+    else:
+        subprocess.Popen(['xdg-open', directory])