+++ libs/utils/source/ftrack_utils/version/__init__.py
@@ -31,3 +31,22 @@ def get_version(package_name, package_path):
             result = toml.load(path_toml)["tool"]["poetry"]["version"]
 
     return result
+
+
+def get_connect_plugin_version(connect_plugin_path):
+    '''Return Connect plugin version string for *connect_plugin_path*'''
+    result = None
+    path_version_file = os.path.join(connect_plugin_path, '__version__.py')
+    if not os.path.isfile(path_version_file):
+        raise FileNotFoundError
+    with open(path_version_file) as f:
+        for line in f.readlines():
+            if line.startswith('__version__'):
+                result = line.split('=')[1].strip().strip("'")
+                break
+    if not result:
+        raise Exception(
+            "Can't extract version number from {}. "
+            "\n Make sure file is valid.".format(path_version_file)
+        )
+    return result