+++ projects/connect-timetracker-widget/release_notes.md
@@ -1,8 +1,9 @@
 # ftrack Connect Timetracker widget release Notes
 
-## v24.2.0
-2024-02-16
+## v24.4.0
+2024-04-02
 
+* [changed] Upgrade ftrack-utils version.
 * [fix] Fixed bug with wrong plugin name reported back with debug information to Connect about dialog.
 * [changed] Ported from ftrack-connect-timetracker-widget repo and aligned with Connect 3
 * [changed] Use poetry as build system.