+++ projects/framework-common-extensions/release_notes.md
@@ -2,6 +2,10 @@
 
 ## Upcoming
 
+* [changed] Asset version selectors; show label on empty asset list.
+* [changed] Publisher dialog only query assets linked to the task.
+* [changed] Opener dialog has checkbox to query assets from AssetBuild
+* [new] Publisher suggest new asset name based on task type.
 * [changed] Add windows specific titles on standard opener dialog and standard publisher dialog.
 * [changed] Added labels to publisher asset selector and publisher, aligned margins.
 * [new] Initial release.