+++ projects/framework-maya/release_notes.md
@@ -1,6 +1,8 @@
 # ftrack Framework Maya integration release Notes
 
-## Upcoming
-YYYY-mm-dd
+## v24.4.0
+2024-04-02
 
+* [changed] maya_export_options_selector widget with all options set to True by default
+* [changed] Upgrade ftrack-framework-core, ftrack-utils and ftrack-qt version.
 * [new] Initial release.