+++ projects/framework-nuke/release_notes.md
@@ -1,6 +1,9 @@
 # ftrack Framework Nuke integration release Notes
 
-## Upcoming
-YYYY-mm-dd
+## v24.4.0
+2024-04-02
 
+* [changed] When selecting the node to render for reviewable, show all nodes with the selected one on top.
+* [new] Writeable node validator, to check if a node is writeable for a reviewable to be created.
+* [changed] Upgrade ftrack-framework-core, ftrack-utils and ftrack-qt version.
 * [new] Initial release.